requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0

# Record linkage and string matching
recordlinkage>=0.14.0
jellyfish>=0.9.0
python-Levenshtein>=0.12.0
faiss-cpu>=1.7.3

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
psutil>=5.9.0
fastdtw>=0.3.4
hmmlearn>=0.3.3

# Visualization and reporting
matplotlib>=3.6.0
seaborn>=0.12.0
jinja2>=3.1.0
plotly>=6.2.0
networkx>=3.5

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Type hints (for development)
typing-extensions>=4.5.0

# Core Deep Learning - CUDA 12.1 enabled
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

mmh3